check number of tuples in the orignal store and the target store before commit; delete from S-Store after a successful migration

+++ src/main/java/istc/bigdawg/migration/FromSStoreToPostgresImplementation.java
@@ -158,11 +158,15 @@ public class FromSStoreToPostgresImplementation implements MigrationImplementati
 	    Long countexportElements = exportTask.get();
 	    Long countLoadedElements = loadTask.get();
 	    
-	    // Delete all tuples from S-Store
-	    String rmTupleStatement = "DELETE FROM " + fromTable;
-	    SStoreSQLHandler sstoreH = new SStoreSQLHandler(connectionFrom);
-	    sstoreH.executeUpdateQuery(rmTupleStatement);
-	    
+	    if (!countexportElements.equals(countLoadedElements)) { // failed
+	    	connectionPostgres.rollback();
+	    } else {
+	    	connectionPostgres.commit();
+	    	// Delete all tuples from S-Store
+	    	String rmTupleStatement = "DELETE FROM " + fromTable;
+	    	SStoreSQLHandler sstoreH = new SStoreSQLHandler(connectionFrom);
+	    	sstoreH.executeUpdateQuery(rmTupleStatement);
+	    }
 
 	    long endTimeMigration = System.currentTimeMillis();
 	    long durationMsec = endTimeMigration - startTimeMigration;
